# docs/swagger.yaml
basePath: /api/v1
definitions:
  model.CostResponse:
    properties:
      total_cost:
        example: 4800
        type: integer
    type: object
  model.CreateSubscriptionRequest:
    properties:
      end_date:
        example: 12-2025
        type: string
      price:
        example: 400
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: 07-2025
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    type: object
  model.ErrorResponse:
    properties:
      error:
        example: something went wrong
        type: string
    type: object
  model.ListSubscriptionsResponse:
    properties:
      subscriptions:
        items:
          $ref: '#/definitions/model.SubscriptionResponse'
        type: array
      total:
        type: integer
    type: object
  model.SubscriptionResponse:
    properties:
      created_at:
        example: "2025-07-01T00:00:00Z"
        type: string
      end_date:
        example: 12-2025
        type: string
      id:
        example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        type: string
      price:
        example: 400
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: 07-2025
        type: string
      updated_at:
        example: "2025-07-01T00:00:00Z"
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    type: object
  model.UpdateSubscriptionRequest:
    properties:
      end_date:
        example: 12-2025
        type: string
      price:
        example: 500
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: 07-2025
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API для агрегации данных об онлайн-подписках пользователей
  title: Subscription Service API
  version: "1.0"
paths:
  /subscriptions:
    get:
      description: Получение списка подписок с пагинацией и фильтрацией
      parameters:
      - description: Фильтр по ID пользователя (UUID)
        in: query
        name: user_id
        type: string
      - description: Фильтр по названию сервиса
        in: query
        name: service_name
        type: string
      - default: 10
        description: Лимит записей
        in: query
        name: limit
        type: integer
      - default: 0
        description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ListSubscriptionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Список подписок
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Создание новой записи о подписке пользователя
      parameters:
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/model.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Создать подписку
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Удаление записи о подписке по ID
      parameters:
      - description: ID подписки (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      description: Получение записи о подписке по её идентификатору
      parameters:
      - description: ID подписки (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Получить подписку по ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Обновление записи о подписке по ID
      parameters:
      - description: ID подписки (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Обновлённые данные
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/model.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Обновить подписку
      tags:
      - subscriptions
  /subscriptions/cost:
    get:
      description: Подсчёт суммарной стоимости всех подписок за выбранный период с
        фильтрацией
      parameters:
      - description: Начало периода (MM-YYYY)
        example: 01-2025
        in: query
        name: period_start
        required: true
        type: string
      - description: Конец периода (MM-YYYY)
        example: 12-2025
        in: query
        name: period_end
        required: true
        type: string
      - description: Фильтр по ID пользователя (UUID)
        in: query
        name: user_id
        type: string
      - description: Фильтр по названию сервиса
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Рассчитать суммарную стоимость подписок
      tags:
      - subscriptions
swagger: "2.0"